add razor pages EF Core for Product

+++ BkpGasProcurementSystem/Models/Product.cs
@@ -13,9 +13,9 @@ namespace BkpGasProcurementSystem.Models
         public string Type { get; set; }
 
         [DataType(DataType.Currency)]
-        public decimal Price { get; set; }
+        public int Price { get; set; }
 
-        public decimal Weight { get; set; }
+        public int Weight { get; set; }
 
     }
 }
